added minutes formatting

+++ TaskScheduler.hpp
@@ -8,5 +8,5 @@ public:
   void Init(void (*func)(), uint32_t interval);
   void UpdateTask(uint32_t current_time);
   uint32_t _interval = 0;
-  uint32_t scratch = 0; // a general purpose variable
+  uint8_t scratch = 0; // a general purpose variable
 };
